Move home widget enum into app module
