--- feature/home/src/test/kotlin/uk/gov/govuk/home/HomeWidgetTest.kt
@@ -1,16 +1,0 @@
-package uk.gov.govuk.home
-
-import junit.framework.TestCase.assertNotNull
-import org.junit.Test
-
-class HomeWidgetTest {
-
-    @Test
-    fun `Given we have a home widget enum class, check it's entries are not null`() {
-        assertNotNull(HomeWidget.FEEDBACK_PROMPT)
-        assertNotNull(HomeWidget.NOTIFICATIONS)
-        assertNotNull(HomeWidget.RECENT_ACTIVITY)
-        assertNotNull(HomeWidget.SEARCH)
-        assertNotNull(HomeWidget.TOPICS)
-    }
-}
